backend/requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-community
chromadb
sentence-transformers
google-generativeai
google-genai
groq
pypdf
python-docx
docx2txt
langdetect
aiosqlite
python-dotenv
python-multipart
pydantic
torch
transformers
requests
beautifulsoup4
